add support for non editable view

+++ app/src/main/java/com/example/binder/ui/viewholder/GroupTypeViewHolder.kt
@@ -1,11 +1,13 @@
 package com.example.binder.ui.viewholder
 
 import android.view.LayoutInflater
+import android.view.View
 import android.view.ViewGroup
 import com.example.binder.databinding.LayoutGroupTypeViewHolderBinding
 import com.example.binder.ui.Item
 import com.example.binder.ui.OnActionListener
 import com.example.binder.ui.viewholder.ViewHolderFactory.Companion.GROUP_CATEGORY_TYPE
+import setVisibility
 
 class GroupTypeViewHolder(parent: ViewGroup, listener: OnActionListener) : BaseViewHolder<Item>(
     listener,
@@ -21,8 +23,12 @@ class GroupTypeViewHolder(parent: ViewGroup, listener: OnActionListener) : BaseV
         (item as? GroupTypeItem)?.let {
             (binding as? LayoutGroupTypeViewHolderBinding)?.let {
                 binding.typeText.text = item.groupType
-                binding.deleteButton.setOnClickListener {
-                    listener.onDeleteRequested(bindingAdapterPosition)
+                if (item.showDeleteButton){
+                    binding.deleteButton.setOnClickListener {
+                        listener.onDeleteRequested(bindingAdapterPosition)
+                    }
+                } else {
+                    binding.deleteButton.visibility = View.GONE
                 }
             }
         }
@@ -36,5 +42,6 @@ class GroupTypeViewHolder(parent: ViewGroup, listener: OnActionListener) : BaseV
 data class GroupTypeItem(
     override val uid: String? = null,
     val groupType:String,
+    val showDeleteButton:Boolean,
     override val type: Int = GROUP_CATEGORY_TYPE
 ): Item()
